Cascading T-Shirt Sizes
Hi George!

My name is justgu.us and I'm an indie dev that loves to create systematical naming conventions that spark the narrative of code in a more semantical way.

I implemented my model for cascading t-shirt sizes on your spaceSteps and spacePairs naming convention, that used to be -2xl etc. It might look fresh at first, but you will soon spot the cascading pattern of the sizing I use; XX, XS, SS, SM, MM, ML, LL, LD, DD, ZZ

It's a way to make fluid scales have a better relationship to each other as tokens. I believe that will improve the user experience of end-users to create their own customPairs like I did, plus it makes for a programmatical pattern to step up or down the scales & sizes if you wish to take it further.

As a bonus, it cleans up the tokens quite nicely and makes them all stick to using the same amount of characters per token.

I added one step that was missing in the current spacePairs; the use of 3XL (-ZZ) as a pair at the end.

Thanks for reviewing and take care!